+++ RaceSync/View Controllers/Standings/StandingsController.swift
@@ -1,0 +1,9 @@
+//
+//  StandingsController.swift
+//  RaceSync
+//
+//  Created by Ignacio Romero Zurbuchen on 2025-12-23.
+//  Copyright © 2025 MultiGP Inc. All rights reserved.
+//
+
+import Foundation